Start spliting into query_os/topology_os and so on
More skeleton code, not every system needs to go into stubs to get num_cores() and
things like that.

For instance, for num_cores() on linux we would either build it from manual
CPUID (coming later this fall on your favorite theater) or parse it from
topology. On Darwin we can get the same from sysctl.