Add amd64 apicid decomposition
This will be the main fallback for x86 and also allows us to build topology for
Darwin and NetBSD.

We can retrieve the apicid from Darwin via IORegistry whatever, we then
decompose and build the thread/core/package relationship.

On NetBSD we can do the setaffinity loop dance, get the apicid and perform the
decomposition.

Most of the C bits I are from openbsd/sys/identifycpu.c that I wrote aeons ago.