Apple's ioreg support and topology for Apple-amd64
Ioreg structure is different from amd64 to arm64 on Apple, this makes it work
for amd64 at least.

Lots of changes and skeleton related to dune and the build system. I'm still not
satisfied with the ioreg dance in lib/, but I've failed to find a better way to
implement this.